api/requirements.txt
asgiref==3.5.2
Babel==2.9.1
bleach==5.0.0
cffi==1.15.1
cli-helpers==2.2.1
Click==7.0
colorama==0.4.5
configobj==5.0.6
cryptography==37.0.4
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.13
django-admin-interface==0.19.0
django-ckeditor==6.4.2
django-colorfield==0.7.0
django-cors-headers==3.12.0
django-crispy-forms==1.14.0
django-extensions==3.2.0
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==2.0.0
django-summernote==0.8.20.0
djangoadmin==2.2.8
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djangotools==2.2.2
filelock==3.6.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
mysqlclient==2.1.0
numpy==1.22.4
opencv-python==4.5.5.64
pendulum==2.1.2
pgcli==3.4.1
pgspecial==1.13.1
Pillow==9.1.1
platformdirs==2.5.1
prompt-toolkit==3.0.30
psycopg2==2.9.3
py3dns==3.2.1
pycparser==2.21
pygame==2.1.2
Pygments==2.12.0
PyJWT==2.3.0
pyOpenSSL==22.0.0
pypm2==0.2.0
python-dateutil==2.8.2
python-emailahoy==0.1
pytz==2022.1
pytzdata==2020.1
setproctitle==1.3.0
six==1.16.0
sqlparse==0.4.2
stdlib-list==0.7.0
tabulate==0.8.10
tzdata==2022.1
validate-email==1.3
virtualenv==20.13.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
whitenoise==6.2.0
